[feature] remove old sql query
