--- src/page/SqlQuery/DatabaseTables/__testData__/index.tsx
@@ -1,39 +1,0 @@
-export const tablesName = [
-  { name: 'audit_plan_report_sqls_v2' },
-  { name: 'audit_plan_reports_v2' },
-  { name: 'audit_plan_sqls_v2' },
-  { name: 'audit_plans' },
-  { name: 'execute_sql_detail' },
-  { name: 'global_configuration_ldap' },
-  { name: 'global_configuration_license' },
-  { name: 'global_configuration_oauth2' },
-  { name: 'global_configuration_smtp' },
-  { name: 'global_configuration_wechat' },
-  { name: 'instance_role' },
-  { name: 'instance_rule_template' },
-  { name: 'instances' },
-  { name: 'role_operations' },
-  { name: 'roles' },
-  { name: 'rollback_sql_detail' },
-  { name: 'rule_template_rule' },
-  { name: 'rule_templates' },
-  { name: 'rules' },
-  { name: 'sql_query_execution_sqls' },
-  { name: 'sql_query_histories' },
-  { name: 'sql_whitelist' },
-  { name: 'system_variables' },
-  { name: 'tasks' },
-  { name: 'user_group_roles' },
-  { name: 'user_group_users' },
-  { name: 'user_groups' },
-  { name: 'user_role' },
-  { name: 'users' },
-  { name: 'workflow_record_history' },
-  { name: 'workflow_records' },
-  { name: 'workflow_step_template_user' },
-  { name: 'workflow_step_templates' },
-  { name: 'workflow_step_user' },
-  { name: 'workflow_steps' },
-  { name: 'workflow_templates' },
-  { name: 'workflows' },
-];
