--- src/page/SqlQuery/DatabaseTables/index.tsx
@@ -1,10 +1,0 @@
-import DatabaseTables from './DatabaseTables';
-
-export type DatabaseTablesProps = {
-  dataSourceName?: string;
-  schemaName?: string;
-
-  getTableSchema: (tableName: string) => Promise<void>;
-};
-
-export default DatabaseTables;
