--- src/page/SqlQuery/SqlInput/Modal/SqlQueryHistory/index.type.ts
@@ -1,10 +1,0 @@
-export interface SqlSearchHistoryProps {
-  visible: boolean;
-  applyHistorySql: (sql: string) => void;
-  close: () => void;
-  instanceName: string;
-}
-
-export type FilterSqlHistoryFormFields = {
-  sql_history_filter: string;
-};
